Updates heading levels and lists on related content
Changing the "Read latest" heading from a `<p>` to a `<h3>`. This makes it consistent
between header and bottom box and it should be a heading.

Removes the `<h4>` from  within the "read latest" box as they should be
links not headings.

Removes the `<h4>` from the  "More of this topic" box as it should be a
list of links not headings.

Put all items in "More on this topic" box into a list.

Changes More on this topic subheadings from `<div>` to `<p>`