Remove duplicate comments tracking (#1459)
Back in the days of iJento, we subscribed to o-comments events and sent tracking calls to record them. In the meantime, tracking has migrated to Spoor, but o-comments has also implemented its own tracking internally. Next-article therefore sends two events to Spoor: `page:comment` and `comment:post`, for example. This is unnecessary.